% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.matur.R
\name{size.matur}
\alias{size.matur}
\title{A nonparametric method to estimate size at first maturity of marine exploited species.}
\usage{
size.matur(size, prop.mat, xlab, ylab, simsize.incr, p.car, perc.size)
}
\arguments{
\item{size}{a vector of size classes (total length, shell length, etc.).}

\item{prop.mat}{Proportion of mature individuals for each size class (must be the same lenght as size).}

\item{xlab}{Character string for the x-axis label of the generated scatter plot (for example total length (cm))}

\item{ylab}{same as xlab but for the y-axis label.}

\item{simsize.incr}{Increase intervals for the simulated vector of sizes, which is used to predict the curve from the fitted GAM. A small interval will create a smoother curve but may take longer time to compute.}

\item{p.car}{A pch value for the marker for the observed maturity data in the generated plot. For example p.car = 19 uses a solid black circle.}

\item{perc.size}{The percentage of the size to estimate the first maturity at. For example, perc.size = 0.50 estimates the size at first maturity of the 50\% of the curve (L50).}
}
\description{
This function fits a Generalized Additive Model (GAM) to size vs maturity (in proportion) data. The function returns a table with the estimated size at first maturity from a percentage of interest of the user (usually 50\%; L50, but also suited for 75\%, 95\%, etc) with the associated 95\% confidence intervals. It also plots the data and the fitted curve with 95\% confidence bands.
}
\examples{
attach(generosa) #geoduck clam data from (Aragon-Noriega 2015)
size.matur(size = generosa$size.mm, prop.mat = generosa$mat.prop, xlab = "Shell size (mm)", ylab = "Proportion of matures", simsize.incr = 0.1, p.car = 19, perc.size = 0.5)
}
\keyword{Marine}
\keyword{Nonparametric}
\keyword{Size}
\keyword{at}
\keyword{exploited}
\keyword{first}
\keyword{maturity;}
\keyword{species}
\keyword{statistics;}
